requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0
pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
opencv-python>=4.5.0
albumentations>=1.3.0
wandb>=0.15.0
tqdm>=4.64.0
accelerate>=0.20.0
datasets>=2.12.0
evaluate>=0.4.0
peft>=0.4.0
bitsandbytes>=0.41.0
google-cloud-aiplatform>=1.35.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0